pawpal/requirements.txt
Django>=4.2
gunicorn
Pillow
django-allauth
requests
asgiref
channels
PyJWT
daphne
promptpay
django-debug-toolbar
